pyproject.toml
[project]
name = "evalobility"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.42.17",
    "botocore>=1.42.17",
    "ddtrace>=4.1.1",
    "dotenv>=0.9.9",
    "strands-agents>=1.20.0",
    "strands-agents-tools>=0.2.18",
]
